// src/api/model/quickScanCommand.ts
/**
 * Receipt Wrangler API.
 * No description provided (generated by Swagger Codegen https://github.com/swagger-api/swagger-codegen)
 *
 * OpenAPI spec version: 0.0.1
 * 
 *
 * NOTE: This class is auto generated by the swagger code generator program.
 * https://github.com/swagger-api/swagger-codegen.git
 * Do not edit the class manually.
 */
import { ReceiptStatus } from './receiptStatus';

export interface QuickScanCommand { 
    file: Blob;
    /**
     * Group foreign key
     */
    groupId: number;
    /**
     * User foreign key
     */
    paidByUserId: number;
    status: ReceiptStatus;
}